changing damage/health to float (warning, changes protocol [rebuild])
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4713 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/CClient.h
@@ -457,8 +457,8 @@ public:
 
 	void		SpawnProjectile(CVec pos, CVec vel, int rot, int owner, proj_t *_proj, int _random, AbsTime time, AbsTime ignoreWormCollBeforeTime);
     void        disableProjectile(CProjectile *prj);
-	void		Explosion(CVec pos, int damage, int shake, int owner);
-	void		InjureWorm(CWorm *w, int damage, int owner);
+	void		Explosion(CVec pos, float damage, int shake, int owner);
+	void		InjureWorm(CWorm *w, float damage, int owner);
 	void		UpdateScoreboard();
 	void		LaserSight(CWorm *w, float Angle, bool highlightCrosshair = true);
